moves WriteBytesTest.FakeFile to tests.actionpack
