+++ tests/actionpack/actions/test_write_bytes.py
@@ -3,10 +3,9 @@ import pickle
 from actionpack import Action
 from actionpack.actions import WriteBytes
 from actionpack.utils import pickleable
+from tests.actionpack import FakeFile
 
 from functools import reduce
-from io import BytesIO
-from io import TextIOWrapper
 from threading import Thread
 from time import sleep
 from unittest import TestCase
@@ -15,29 +14,6 @@ from unittest.mock import patch
 
 class WriteBytesTest(TestCase):
 
-    class FakeFile:
-        def __init__(self, contents: bytes=bytes(), mode: str=None):
-            self.buffer = BytesIO(contents)
-            self.buffer.read()
-            self.mode = mode
-
-        def read(self):
-            self.buffer.seek(0)
-            return self.buffer.read()
-
-        def write(self, data: bytes):
-            if self.mode == 'wb':
-                self.buffer.seek(0)
-                self.buffer.truncate()
-            self.buffer.write(data)
-            return len(data)
-
-        def __enter__(self):
-            return self
-
-        def __exit__(self, *args):
-            self.closed = True
-
     def setUp(self):
         self.salutation = 'Hello.'.encode()
         self.question = ' How are you?'.encode()
@@ -45,7 +21,7 @@ class WriteBytesTest(TestCase):
 
     @patch('pathlib.Path.open')
     def test_can_WriteBytes(self, mock_output):
-        file = self.FakeFile(self.salutation)
+        file = FakeFile(self.salutation)
         mock_output.return_value = file
         result = self.action.perform()
 
@@ -54,7 +30,7 @@ class WriteBytesTest(TestCase):
 
     @patch('pathlib.Path.open')
     def test_can_overWriteBytes(self, mock_output):
-        file = self.FakeFile(self.salutation, 'wb')
+        file = FakeFile(self.salutation, 'wb')
         mock_output.return_value = file
         action = WriteBytes('valid/path/to/file', self.question, overwrite=True)
         result = self.action.perform()
@@ -64,7 +40,7 @@ class WriteBytesTest(TestCase):
 
     @patch('pathlib.Path.open')
     def test_can_WriteBytes_concurrently(self, mock_output):
-        file = self.FakeFile(self.salutation, 'wb')
+        file = FakeFile(self.salutation, 'wb')
         mock_output.return_value = file
         initial_file_contents = file.read()
         filepath = 'valid/path/to/file'
